RAG stablization, Dynamic filter on DF to feed context query
