+++ data/last_update_date.txt
@@ -1,1 +1,1 @@
-2025-11-10
+2025-11-18
